+++ src/realm/sync/tools/apply_to_state_command.cpp
@@ -3,12 +3,12 @@
 #include "realm/sync/instruction_applier.hpp"
 #include "realm/sync/impl/clamped_hex_dump.hpp"
 #include "realm/sync/noinst/client_history_impl.hpp"
+#include "realm/sync/noinst/protocol_codec.hpp"
 #include "realm/sync/protocol.hpp"
 #include "realm/sync/transform.hpp"
 #include "realm/sync/changeset_parser.hpp"
 #include "realm/sync/noinst/compression.hpp"
 #include "realm/util/cli_args.hpp"
-#include "realm/util/from_chars.hpp"
 #include "realm/util/load_file.hpp"
 #include "realm/util/safe_int_ops.hpp"
 #include "realm/util/string_view.hpp"
@@ -20,17 +20,16 @@
 #include <string>
 #include <type_traits>
 
-
+namespace {
+using namespace realm;
 using namespace realm::util;
-
-template <typename T>
-using ParseResult = std::pair<T, StringView>;
+using namespace realm::_impl;
 
 struct ServerIdentMessage {
     realm::sync::session_ident_type session_ident;
     realm::sync::SaltedFileIdent file_ident;
 
-    static ParseResult<ServerIdentMessage> parse(StringView sv);
+    static ServerIdentMessage parse(HeaderLineParser& msg);
 };
 
 struct DownloadMessage {
@@ -42,7 +41,7 @@ struct DownloadMessage {
     Buffer<char> uncompressed_body_buffer;
     std::vector<realm::sync::Transformer::RemoteChangeset> changesets;
 
-    static ParseResult<DownloadMessage> parse(StringView sv, Logger& logger);
+    static DownloadMessage parse(HeaderLineParser& msg, Logger& logger);
 };
 
 struct UploadMessage {
@@ -53,157 +52,51 @@ struct UploadMessage {
     Buffer<char> uncompressed_body_buffer;
     std::vector<realm::sync::Changeset> changesets;
 
-    static ParseResult<UploadMessage> parse(StringView sv, Logger& logger);
+    static UploadMessage parse(HeaderLineParser& msg, Logger& logger);
 };
 
 using Message = mpark::variant<ServerIdentMessage, DownloadMessage, UploadMessage>;
 
-struct MessageParseException : public std::runtime_error {
-    using std::runtime_error::runtime_error;
-
-    template <typename... Args>
-    MessageParseException(const char* fmt, Args&&... args)
-        : std::runtime_error(format(fmt, args...))
-    {
-    }
-};
-
-// These functions will parse the space/new-line delimited headers found at the beginning of
-// messages and changesets.
-StringView parse_header_element(StringView sv, char)
+Message parse_message(HeaderLineParser& msg, Logger& logger)
 {
-    return sv;
+    auto message_type = msg.read_next<std::string_view>();
+    if (message_type == "download") {
+        return DownloadMessage::parse(msg, logger);
+    }
+    else if (message_type == "upload") {
+        return UploadMessage::parse(msg, logger);
+    }
+    else if (message_type == "ident") {
+        return ServerIdentMessage::parse(msg);
+    }
+    throw ProtocolCodecException("could not find valid message in input");
 }
 
-template <typename T, typename = std::enable_if_t<std::is_integral_v<std::remove_reference_t<T>>>>
-StringView parse_header_value(StringView sv, T&& cur_arg)
+ServerIdentMessage ServerIdentMessage::parse(HeaderLineParser& msg)
 {
-    auto parse_res = realm::util::from_chars(sv.begin(), sv.end(), cur_arg, 10);
-    if (parse_res.ec != std::errc{}) {
-        throw MessageParseException("error parsing integer in header line: %1",
-                                    std::make_error_code(parse_res.ec).message());
-    }
-
-    return sv.substr(parse_res.ptr - sv.begin());
-}
-
-StringView parse_header_value(StringView sv, StringView& cur_arg)
-{
-    auto delim_at = std::find(sv.begin(), sv.end(), ' ');
-    if (delim_at == sv.end()) {
-        throw MessageParseException("reached end of header line prematurely");
-    }
-
-    auto sub_str_len = std::distance(sv.begin(), delim_at);
-    cur_arg = StringView(sv.begin(), sub_str_len);
-
-    return sv.substr(sub_str_len);
-}
-
-template <typename T, typename... Args>
-StringView parse_header_element(StringView sv, char end_delim, T&& cur_arg, Args&&... next_args)
-{
-    if (sv.empty()) {
-        throw MessageParseException("cannot parse an empty header line");
-    }
-    sv = parse_header_value(sv, std::forward<T&&>(cur_arg));
-
-    if (sv.front() == ' ') {
-        return parse_header_element(sv.substr(1), end_delim, next_args...);
-    }
-    if (sv.front() == end_delim) {
-        return sv.substr(1);
-    }
-    throw MessageParseException("found invalid character in header line");
-}
-
-template <typename... Args>
-StringView parse_header_line(StringView sv, char end_delim, Args&&... args)
-{
-    return parse_header_element(sv, end_delim, args...);
-}
-
-struct MessageBody {
-    StringView body_view;
-    StringView remaining;
-    Buffer<char> uncompressed_body_buffer;
-    static MessageBody parse(StringView sv, std::size_t compressed_body_size, std::size_t uncompressed_body_size,
-                             bool is_body_compressed);
-};
-
-MessageBody MessageBody::parse(StringView sv, std::size_t compressed_body_size, std::size_t uncompressed_body_size,
-                               bool is_body_compressed)
-{
-    MessageBody ret;
-    if (is_body_compressed) {
-        if (sv.size() < compressed_body_size) {
-            throw MessageParseException("compressed message body is bigger (%1) than available bytes (%2)",
-                                        compressed_body_size, sv.size());
-        }
-
-        ret.uncompressed_body_buffer.set_size(uncompressed_body_size);
-        auto ec = realm::_impl::compression::decompress(sv.data(), compressed_body_size,
-                                                        ret.uncompressed_body_buffer.data(), uncompressed_body_size);
-        if (ec) {
-            throw MessageParseException("error decompressing message body: %1", ec.message());
-        }
-
-        ret.remaining = sv.substr(compressed_body_size);
-        ret.body_view = StringView{ret.uncompressed_body_buffer.data(), uncompressed_body_size};
-    }
-    else {
-        if (sv.size() < uncompressed_body_size) {
-            throw MessageParseException("message body is bigger (%1) than available bytes (%2)",
-                                        uncompressed_body_size, sv.size());
-        }
-        ret.body_view = sv.substr(0, uncompressed_body_size);
-        ret.remaining = sv.substr(uncompressed_body_size);
-    }
+    ServerIdentMessage ret;
+    ret.session_ident = msg.read_next<sync::session_ident_type>();
+    ret.file_ident.ident = msg.read_next<sync::file_ident_type>();
+    ret.file_ident.salt = msg.read_next<sync::salt_type>('\n');
 
     return ret;
 }
 
-ParseResult<Message> parse_message(StringView sv, Logger& logger)
-{
-    StringView message_type;
-    sv = parse_header_element(sv, ' ', message_type);
-
-    if (message_type == "download") {
-        return DownloadMessage::parse(sv, logger);
-    }
-    else if (message_type == "upload") {
-        return UploadMessage::parse(sv, logger);
-    }
-    else if (message_type == "ident") {
-        return ServerIdentMessage::parse(sv);
-    }
-    throw MessageParseException("could not find valid message in input");
-}
-
-ParseResult<ServerIdentMessage> ServerIdentMessage::parse(StringView sv)
-{
-    ServerIdentMessage ret;
-    sv = parse_header_line(sv, '\n', ret.session_ident, ret.file_ident.ident, ret.file_ident.salt);
-
-    return std::make_pair(std::move(ret), sv);
-}
-
-ParseResult<DownloadMessage> DownloadMessage::parse(StringView sv, Logger& logger)
+DownloadMessage DownloadMessage::parse(HeaderLineParser& msg, Logger& logger)
 {
     DownloadMessage ret;
-    int is_body_compressed;
-    std::size_t uncompressed_body_size, compressed_body_size;
 
-    sv = parse_header_line(sv, '\n', ret.session_ident, ret.progress.download.server_version,
-                           ret.progress.download.last_integrated_client_version, ret.latest_server_version.version,
-                           ret.latest_server_version.salt, ret.progress.upload.client_version,
-                           ret.progress.upload.last_integrated_server_version, ret.downloadable_bytes,
-                           is_body_compressed, uncompressed_body_size, compressed_body_size);
-
-    auto message_body = MessageBody::parse(sv, compressed_body_size, uncompressed_body_size, is_body_compressed);
-    ret.uncompressed_body_buffer = std::move(message_body.uncompressed_body_buffer);
-    sv = message_body.remaining;
-    auto body_view = message_body.body_view;
+    ret.session_ident = msg.read_next<sync::session_ident_type>();
+    ret.progress.download.server_version = msg.read_next<sync::version_type>();
+    ret.progress.download.last_integrated_client_version = msg.read_next<sync::version_type>();
+    ret.progress.latest_server_version.version = msg.read_next<sync::version_type>();
+    ret.progress.latest_server_version.salt = msg.read_next<sync::salt_type>();
+    ret.progress.upload.client_version = msg.read_next<sync::version_type>();
+    ret.progress.upload.last_integrated_server_version = msg.read_next<sync::version_type>();
+    ret.downloadable_bytes = msg.read_next<int64_t>();
+    auto is_body_compressed = msg.read_next<bool>();
+    auto uncompressed_body_size = msg.read_next<size_t>();
+    auto compressed_body_size = msg.read_next<size_t>('\n');
 
     logger.trace("decoding download message. "
                  "{download: {server: %1, client: %2} upload: {server: %3, client: %4}, latest: %5}",
@@ -211,63 +104,93 @@ ParseResult<DownloadMessage> DownloadMessage::parse(StringView sv, Logger& logge
                  ret.progress.upload.last_integrated_server_version, ret.progress.upload.client_version,
                  ret.latest_server_version.version);
 
-    while (!body_view.empty()) {
-        realm::sync::Transformer::RemoteChangeset cur_changeset;
-        std::size_t changeset_size;
-        body_view =
-            parse_header_line(body_view, ' ', cur_changeset.remote_version,
-                              cur_changeset.last_integrated_local_version, cur_changeset.origin_timestamp,
-                              cur_changeset.origin_file_ident, cur_changeset.original_changeset_size, changeset_size);
-        if (changeset_size > body_view.size()) {
-            throw MessageParseException("changeset length is %1 but buffer size is %2", changeset_size,
-                                        body_view.size());
+    std::unique_ptr<char[]> uncompressed_body_buffer;
+    HeaderLineParser body;
+    if (is_body_compressed) {
+        uncompressed_body_buffer = std::make_unique<char[]>(uncompressed_body_size);
+        auto compressed_body = msg.read_sized_data<BinaryData>(compressed_body_size);
+        std::error_code ec = _impl::compression::decompress(compressed_body.data(), compressed_body.size(),
+                                                            uncompressed_body_buffer.get(), uncompressed_body_size);
+
+        if (ec) {
+            throw ProtocolCodecException("error decompressing download message");
         }
 
+        body = HeaderLineParser(std::string_view(uncompressed_body_buffer.get(), uncompressed_body_size));
+    }
+    else {
+        body = HeaderLineParser(msg.read_sized_data<std::string_view>(uncompressed_body_size));
+    }
+
+    while (!body.at_end()) {
+        realm::sync::Transformer::RemoteChangeset cur_changeset;
+        cur_changeset.remote_version = body.read_next<sync::version_type>();
+        cur_changeset.last_integrated_local_version = body.read_next<sync::version_type>();
+        cur_changeset.origin_timestamp = body.read_next<sync::timestamp_type>();
+        cur_changeset.origin_file_ident = body.read_next<sync::file_ident_type>();
+        cur_changeset.original_changeset_size = body.read_next<size_t>();
+        auto changeset_size = body.read_next<size_t>();
+
         realm::sync::Changeset parsed_changeset;
-        realm::_impl::SimpleNoCopyInputStream changeset_stream(body_view.data(), changeset_size);
+        auto changeset_data = body.read_sized_data<BinaryData>(changeset_size);
+        auto changeset_stream = realm::_impl::SimpleNoCopyInputStream(changeset_data.data(), changeset_data.size());
         realm::sync::parse_changeset(changeset_stream, parsed_changeset);
         logger.trace("found download changeset: serverVersion: %1, clientVersion: %2, origin: %3 %4",
                      cur_changeset.remote_version, cur_changeset.last_integrated_local_version,
                      cur_changeset.origin_file_ident, parsed_changeset);
-        realm::BinaryData changeset_data{body_view.data(), changeset_size};
         cur_changeset.data = changeset_data;
         ret.changesets.push_back(cur_changeset);
-        body_view = body_view.substr(changeset_size);
     }
 
-    return std::make_pair(std::move(ret), sv);
+    return ret;
 }
 
-ParseResult<UploadMessage> UploadMessage::parse(StringView sv, Logger& logger)
+UploadMessage UploadMessage::parse(HeaderLineParser& msg, Logger& logger)
 {
     UploadMessage ret;
-    int is_body_compressed;
-    std::size_t uncompressed_body_size, compressed_body_size;
 
-    sv = parse_header_line(sv, '\n', ret.session_ident, is_body_compressed, uncompressed_body_size,
-                           compressed_body_size, ret.upload_progress.client_version,
-                           ret.upload_progress.last_integrated_server_version, ret.locked_server_version);
+    ret.session_ident = msg.read_next<sync::session_ident_type>();
+    auto is_body_compressed = msg.read_next<bool>();
+    auto uncompressed_body_size = msg.read_next<size_t>();
+    auto compressed_body_size = msg.read_next<size_t>();
+    ret.upload_progress.client_version = msg.read_next<sync::version_type>();
+    ret.upload_progress.last_integrated_server_version = msg.read_next<sync::version_type>();
+    ret.locked_server_version = msg.read_next<sync::version_type>('\n');
 
-    auto message_body = MessageBody::parse(sv, compressed_body_size, uncompressed_body_size, is_body_compressed);
-    ret.uncompressed_body_buffer = std::move(message_body.uncompressed_body_buffer);
-    sv = message_body.remaining;
-    auto body_view = message_body.body_view;
+    std::unique_ptr<char[]> uncompressed_body_buffer;
+    HeaderLineParser body;
+    // if is_body_compressed == true, we must decompress the received body.
+    if (is_body_compressed) {
+        uncompressed_body_buffer = std::make_unique<char[]>(uncompressed_body_size);
+        auto compressed_body = msg.read_sized_data<BinaryData>(compressed_body_size);
 
-    while (!body_view.empty()) {
-        realm::sync::Changeset cur_changeset;
-        std::size_t changeset_size;
-        body_view =
-            parse_header_line(body_view, ' ', cur_changeset.version, cur_changeset.last_integrated_remote_version,
-                              cur_changeset.origin_timestamp, cur_changeset.origin_file_ident, changeset_size);
-        if (changeset_size > body_view.size()) {
-            throw MessageParseException("changeset length in upload message is %1 but bufer size is %2",
-                                        changeset_size, body_view.size());
+        std::error_code ec = _impl::compression::decompress(compressed_body.data(), compressed_body.size(),
+                                                            uncompressed_body_buffer.get(), uncompressed_body_size);
+
+        if (ec) {
+            throw ProtocolCodecException("error decompressing upload message");
         }
+
+        body = HeaderLineParser(std::string_view(uncompressed_body_buffer.get(), uncompressed_body_size));
+    }
+    else {
+        body = HeaderLineParser(msg.read_sized_data<std::string_view>(uncompressed_body_size));
+    }
+
+    while (!body.at_end()) {
+        realm::sync::Changeset cur_changeset;
+        cur_changeset.version = body.read_next<sync::version_type>();
+        cur_changeset.last_integrated_remote_version = body.read_next<sync::version_type>();
+        cur_changeset.origin_timestamp = body.read_next<sync::timestamp_type>();
+        cur_changeset.origin_file_ident = body.read_next<sync::file_ident_type>();
+        auto changeset_size = body.read_next<size_t>();
+
+        auto changeset_buffer = body.read_sized_data<BinaryData>(changeset_size);
 
         logger.trace("found upload changeset: %1 %2 %3 %4 %5", cur_changeset.last_integrated_remote_version,
                      cur_changeset.version, cur_changeset.origin_timestamp, cur_changeset.origin_file_ident,
                      changeset_size);
-        realm::_impl::SimpleNoCopyInputStream changeset_stream(body_view.data(), changeset_size);
+        realm::_impl::SimpleNoCopyInputStream changeset_stream(changeset_buffer.data(), changeset_buffer.size());
         try {
             realm::sync::parse_changeset(changeset_stream, cur_changeset);
         }
@@ -277,10 +200,9 @@ ParseResult<UploadMessage> UploadMessage::parse(StringView sv, Logger& logger)
         }
         logger.trace("Decoded changeset: %1", cur_changeset);
         ret.changesets.push_back(std::move(cur_changeset));
-        body_view = body_view.substr(changeset_size);
     }
 
-    return std::make_pair(std::move(ret), sv);
+    return ret;
 }
 
 void print_usage(StringView program_name)
@@ -302,6 +224,8 @@ void print_usage(StringView program_name)
                  "                       command belongs to."
               << std::endl;
 }
+
+} // namespace
 
 int main(int argc, const char** argv)
 {
@@ -348,18 +272,17 @@ int main(int argc, const char** argv)
     auto local_db = realm::DB::create(history, db_opts);
 
     auto input_contents = load_file(input_arg.as<std::string>());
-    auto input_view = StringView{input_contents};
-    while (!input_view.empty()) {
-        ParseResult<Message> message;
+    HeaderLineParser msg(input_contents);
+    while (!msg.at_end()) {
+        Message message;
         try {
-            message = parse_message(input_view, *logger);
+            message = parse_message(msg, *logger);
         }
-        catch (const MessageParseException& e) {
+        catch (const ProtocolCodecException& e) {
             logger->error("Error parsing input message file: %1", e.what());
             return EXIT_FAILURE;
         }
 
-        input_view = message.second;
         bool download_integration_failed = false;
         mpark::visit(realm::util::overload{
                          [&](const DownloadMessage& download_message) {
@@ -389,7 +312,7 @@ int main(int argc, const char** argv)
                          [&](const ServerIdentMessage& ident_message) {
                              history.set_client_file_ident(ident_message.file_ident, true);
                          }},
-                     message.first);
+                     message);
         if (download_integration_failed) {
             return EXIT_FAILURE;
         }